meta: Extract revset parser as a separate crate
