+++ lib/src/revset.rs
@@ -28,6 +28,9 @@ use futures::Stream;
 use futures::StreamExt as _;
 use futures::stream::LocalBoxStream;
 use itertools::Itertools as _;
+pub use jj_revset_parser::fmt::format_remote_symbol;
+pub use jj_revset_parser::fmt::format_string;
+pub use jj_revset_parser::fmt::format_symbol;
 use pollster::FutureExt as _;
 use thiserror::Error;
 
@@ -3540,31 +3543,6 @@ pub struct RevsetWorkspaceContext<'a> {
     pub workspace_name: &'a WorkspaceName,
 }
 
-/// Formats a string as symbol by quoting and escaping it if necessary.
-///
-/// Note that symbols may be substituted to user aliases. Use
-/// [`format_string()`] to ensure that the provided string is resolved as a
-/// tag/bookmark name, commit/change ID prefix, etc.
-pub fn format_symbol(literal: &str) -> String {
-    if revset_parser::is_identifier(literal) {
-        literal.to_string()
-    } else {
-        format_string(literal)
-    }
-}
-
-/// Formats a string by quoting and escaping it.
-pub fn format_string(literal: &str) -> String {
-    format!(r#""{}""#, dsl_util::escape_string(literal))
-}
-
-/// Formats a `name@remote` symbol, applies quoting and escaping if necessary.
-pub fn format_remote_symbol(name: &str, remote: &str) -> String {
-    let name = format_symbol(name);
-    let remote = format_symbol(remote);
-    format!("{name}@{remote}")
-}
-
 #[cfg(test)]
 #[rustversion::attr(
     since(1.89),
@@ -6394,41 +6372,5 @@ mod tests {
         }
         "#);
         Ok(())
-    }
-
-    #[test]
-    fn test_escape_string_literal() {
-        // Valid identifiers don't need quoting
-        assert_eq!(format_symbol("foo"), "foo");
-        assert_eq!(format_symbol("foo.bar"), "foo.bar");
-
-        // Invalid identifiers need quoting
-        assert_eq!(format_symbol("foo@bar"), r#""foo@bar""#);
-        assert_eq!(format_symbol("foo bar"), r#""foo bar""#);
-        assert_eq!(format_symbol(" foo "), r#"" foo ""#);
-        assert_eq!(format_symbol("(foo)"), r#""(foo)""#);
-        assert_eq!(format_symbol("all:foo"), r#""all:foo""#);
-
-        // Some characters also need escaping
-        assert_eq!(format_symbol("foo\"bar"), r#""foo\"bar""#);
-        assert_eq!(format_symbol("foo\\bar"), r#""foo\\bar""#);
-        assert_eq!(format_symbol("foo\\\"bar"), r#""foo\\\"bar""#);
-        assert_eq!(format_symbol("foo\nbar"), r#""foo\nbar""#);
-
-        // Some characters don't technically need escaping, but we escape them for
-        // clarity
-        assert_eq!(format_symbol("foo\"bar"), r#""foo\"bar""#);
-        assert_eq!(format_symbol("foo\\bar"), r#""foo\\bar""#);
-        assert_eq!(format_symbol("foo\\\"bar"), r#""foo\\\"bar""#);
-        assert_eq!(format_symbol("foo \x01 bar"), r#""foo \x01 bar""#);
-    }
-
-    #[test]
-    fn test_escape_remote_symbol() {
-        assert_eq!(format_remote_symbol("foo", "bar"), "foo@bar");
-        assert_eq!(
-            format_remote_symbol(" foo ", "bar:baz"),
-            r#"" foo "@"bar:baz""#
-        );
     }
 }
